Revise getting started topic with latest changes
I revised the topic to reflect the latest changes that were made based on the TOC revisions. I suppose there is a chance it might need further revision, based on feedback, but I wanted to get it started. (ha) I am not sure about the toctree, nor the comment I inserted. Revise as necessary.